requirements.txt
meya-sdk==2.0.0
meya-cli==2.0.1
